Include useragent addition
It would be nice to  also include a build ID or version number in the user
agent, but that's kind of a yak shave with the current build process. (Because
we rely heavily on the logspout Dockerfile, which doesn't give us a good way to
inject a build ID.) So punting on that for now.